refactor: Simplify import paths by creating index.ts files in common and home directories
