+++ src/components/common/GoodsItem/Ranking/index.stories.tsx
@@ -1,10 +1,15 @@
 import { Meta, StoryObj } from '@storybook/react';
-import GoodsItem, { GoodsItemProps } from '@components/common/GoodsItem/GoodItem';
+import GoodsItem, { GoodsItemProps } from '@/components/common/GoodsItem';
 
 const meta: Meta<GoodsItemProps> = {
-  title: 'common/GoodsItem/Default',
-  tags: ['autodocs'],
+  title: 'common/GoodsItem/Ranking',
   component: GoodsItem,
+  tags: ['autodocs'],
+  argTypes: {
+    rankingIndex: {
+      control: { type: 'number', min: 1, max: 10 },
+    },
+  },
 };
 
 export default meta;
@@ -17,5 +22,6 @@ export const Default: Story = {
     subtitle: '부제목',
     title: '상품 제목',
     amount: 10000,
+    rankingIndex: 1,
   },
 };
