+++ src/components/common/GoodsItem/Ranking/index.tsx
