+++ src/components/common/index.ts
@@ -1,0 +1,10 @@
+export { default as Button } from './Button';
+export { default as Footer } from './Footer';
+export { default as Input } from './Form/Input';
+export { default as Image } from './Image';
+export { default as GoodsItem } from './GoodsItem';
+export { default as Ranking } from './GoodsItem/Ranking';
+export { default as Header } from './Header';
+export { default as Container } from './Layout/Container';
+export { default as Grid } from './Layout/Grid';
+export { default as Inner } from './Layout/Inner';
