+++ src/components/home/index.ts
@@ -1,0 +1,5 @@
+export { default as Filter } from './Filter';
+export { default as FriendSelector } from './FriendSelector';
+export { default as RankingList } from './RankingList';
+export { default as ThemeCategory } from './ThemeCategory';
+export { default as TrendingGifts } from './TrendingGifts';
