+++ src/pages/Home.tsx
@@ -1,9 +1,6 @@
 import React from 'react';
-import Header from '@components/common/Header';
-import Footer from '@components/common/Footer';
-import FriendSelector from '@components/home/FriendSelector';
-import ThemeCategory from '@components/home/ThemeCategory';
-import TrendingGifts from '@components/home/TrendingGifts';
+import { Header, Footer } from '@components/common';
+import { FriendSelector, ThemeCategory, TrendingGifts } from '@components/home';
 
 export default function Home() {
   return (
